Dongle i18n: translate via the existing auto-translate-arb plugin (#460)
Reuse the project's existing translation machinery instead of a bespoke path:
the dongle's mail + web-UI strings now live as ARB and are translated at
development time by de.haumacher:auto-translate-arb — the SAME Gradle plugin
phoneblock_mobile uses, with the same `deepl` server credential and per-key
CRC reuse-cache. Outputs are committed; nothing is translated at release time.

- scripts/i18n/l10n/: a standalone Gradle project (settings.gradle + mail/ and
  ui/ subprojects) applying the plugin. Sources:
  - l10n/mail/mail_de.arb  (from the former mail.de.json; the German mail source)
  - l10n/ui/ui_de.arb      (regenerated from index.html's inline I18N.de)
  Translated l10n/**/*_{en,fr,es}.arb are committed (en/fr/es UI seeded verbatim
  from the previously-extracted reviewed packs — byte-identical after strip).
- gen-ui-de-arb.js: regenerates ui_de.arb from index.html, preserving the
  plugin's x-translated CRC for unchanged keys so only changed strings re-run.
- i18n-assets.sh: consumes the committed ARB, stripping @key/@@locale metadata
  into the clean published packs (mail-<lang>.json, ui/lang-<code>.json).
- Removed the interim flat packs (mail.de.json, ui/lang-*.json) and the
  extract-de-ui.js helper (superseded by gen-ui-de-arb.js).

Verified behaviour-preserving: strip(mail_de.arb) == old mail.de.json and
strip(ui_en.arb) == old lang-en.json (byte-identical), so the firmware/browser
see exactly the same strings. Firmware builds; host suite 8/8; signing
roundtrip 8/8; release dry-run assembles + signs from the ARB sources with the
firmware-equivalent signature verify passing.

Dev flow to add/refresh translations:
  node scripts/i18n/gen-ui-de-arb.js         # if German UI changed
  cd scripts/i18n/l10n && gradle translateArb
  git add scripts/i18n/l10n                   # commit the updated *_<lang>.arb
