+++ phoneblock-dongle/firmware/scripts/i18n/l10n/mail/mail_de.arb
@@ -1,0 +1,131 @@
+{
+  "@@locale": "de",
+  "intro": "Statusmeldung deines PhoneBlock-Dongles.",
+  "@intro": {
+    "description": "Status mail: intro line"
+  },
+  "subj.error_and_calls": "PhoneBlock-Dongle: Fehler und neue Anrufe",
+  "@subj.error_and_calls": {
+    "description": "Status-mail string (subj.error_and_calls); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "subj.error": "PhoneBlock-Dongle: Fehler im Protokoll",
+  "@subj.error": {
+    "description": "Status-mail string (subj.error); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "subj.calls": "PhoneBlock-Dongle: Neue Anrufe",
+  "@subj.calls": {
+    "description": "Status-mail string (subj.calls); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "subj.status": "PhoneBlock-Dongle: Statusmeldung",
+  "@subj.status": {
+    "description": "Status-mail string (subj.status); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "subj.update": "PhoneBlock-Dongle: Firmware aktualisiert",
+  "@subj.update": {
+    "description": "Status-mail string (subj.update); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "newcalls.one": "Seit der letzten Meldung ist <b>1</b> neuer Anruf eingegangen.",
+  "@newcalls.one": {
+    "description": "Status-mail string (newcalls.one); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "newcalls.many": "Seit der letzten Meldung sind <b>%d</b> neue Anrufe eingegangen.",
+  "@newcalls.many": {
+    "description": "Status-mail string (newcalls.many); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "sum.device": "Ger&auml;t:",
+  "@sum.device": {
+    "description": "Summary label: device"
+  },
+  "sum.uptime": "Laufzeit: %lldh %lldmin",
+  "@sum.uptime": {
+    "description": "Summary: uptime (%lld hours/minutes)"
+  },
+  "sum.calls": "Anrufe gesamt: %u | SPAM blockiert: %u | durchgestellt: %u",
+  "@sum.calls": {
+    "description": "Summary: call counters (%u total/blocked/passed)"
+  },
+  "calls.heading": "Letzte Anrufe",
+  "@calls.heading": {
+    "description": "Recent-calls table heading"
+  },
+  "calls.time": "Zeit",
+  "@calls.time": {
+    "description": "Table header: time"
+  },
+  "calls.number": "Nummer",
+  "@calls.number": {
+    "description": "Table header: number"
+  },
+  "calls.name": "Name",
+  "@calls.name": {
+    "description": "Table header: name"
+  },
+  "calls.rating": "Bewertung",
+  "@calls.rating": {
+    "description": "Table header: rating"
+  },
+  "verdict.scope.range": "Bereich",
+  "@verdict.scope.range": {
+    "description": "Status-mail string (verdict.scope.range); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.scope.number": "Nummer",
+  "@verdict.scope.number": {
+    "description": "Status-mail string (verdict.scope.number); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.spam_blacklist": "SPAM (Blacklist, %s)",
+  "@verdict.spam_blacklist": {
+    "description": "Status-mail string (verdict.spam_blacklist); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.spam_blocklist": "SPAM (Blockliste, %s)",
+  "@verdict.spam_blocklist": {
+    "description": "Status-mail string (verdict.spam_blocklist); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.spam": "SPAM",
+  "@verdict.spam": {
+    "description": "Status-mail string (verdict.spam); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.spam_votes": "SPAM (%d direkt, %d Range)",
+  "@verdict.spam_votes": {
+    "description": "Status-mail string (verdict.spam_votes); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.spam_suspect": "SPAM-VERDACHT (%d direkt, %d Range)",
+  "@verdict.spam_suspect": {
+    "description": "Status-mail string (verdict.spam_suspect); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.legitimate": "legitim",
+  "@verdict.legitimate": {
+    "description": "Status-mail string (verdict.legitimate); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.error": "Fehler",
+  "@verdict.error": {
+    "description": "Status-mail string (verdict.error); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "verdict.unknown": "unbekannt",
+  "@verdict.unknown": {
+    "description": "Status-mail string (verdict.unknown); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "log.heading": "Neue Meldungen im Protokoll",
+  "@log.heading": {
+    "description": "Log-section heading"
+  },
+  "update.intro1": "Die Firmware auf deinem ",
+  "@update.intro1": {
+    "description": "Status-mail string (update.intro1); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "update.link": "PhoneBlock-Dongle",
+  "@update.link": {
+    "description": "Status-mail string (update.link); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "update.intro2": " wurde auf ",
+  "@update.intro2": {
+    "description": "Status-mail string (update.intro2); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "update.version_prefix": "Version ",
+  "@update.version_prefix": {
+    "description": "Status-mail string (update.version_prefix); keep %d/%s/%u/%lld and <b> tags"
+  },
+  "update.intro3": " aktualisiert.",
+  "@update.intro3": {
+    "description": "Status-mail string (update.intro3); keep %d/%s/%u/%lld and <b> tags"
+  }
+}
