+++ phoneblock-dongle/firmware/scripts/i18n/l10n/settings.gradle
@@ -1,0 +1,18 @@
+// Standalone Gradle project that translates the dongle's German string
+// sources into the other locales using the SAME plugin the mobile app uses
+// (de.haumacher:auto-translate-arb), at development time. Its outputs
+// (mail_<lang>.arb, ui_<lang>.arb) are committed to git; the release script
+// only assembles those committed files — no translation runs at release time.
+//
+// Run:  cd scripts/i18n/l10n && gradle translateArb
+//       (needs a DeepL key under the `deepl` server id in ~/.m2/settings.xml,
+//        exactly like phoneblock_mobile's translateArb and the web plugins.)
+pluginManagement {
+    repositories {
+        mavenLocal()
+        gradlePluginPortal()
+        mavenCentral()
+    }
+}
+rootProject.name = 'dongle-i18n'
+include 'mail', 'ui'
