+++ phoneblock-dongle/firmware/scripts/i18n/l10n/ui/build.gradle
@@ -1,0 +1,14 @@
+// Translate the dongle web-UI strings (German source) into the other
+// locales. Same plugin/credential as phoneblock_mobile's translateArb.
+//
+// ui_de.arb is regenerated from the inline `I18N.de` in main/web/index.html
+// (the runtime German source) by ../../gen-ui-de-arb.js — run that before
+// translating whenever the German UI text changes, then `gradle translateArb`.
+plugins {
+    id 'de.haumacher.auto-translate-arb' version '1.1.0'
+}
+translateArb {
+    serverId    = 'deepl'
+    sourceFile  = file('ui_de.arb')
+    targetLangs = ['en', 'fr', 'es']
+}
