+++ phoneblock-dongle/firmware/scripts/i18n/l10n/ui/ui_en.arb
@@ -1,4 +1,5 @@
 {
+  "@@locale": "en",
   "nav.home": "Status",
   "status.title": "Status",
   "status.sip.connected": "Registered",
