+++ phoneblock-dongle/firmware/scripts/i18n/l10n/ui/ui_es.arb
@@ -1,4 +1,5 @@
 {
+  "@@locale": "es",
   "nav.home": "Estado",
   "status.title": "Estado",
   "status.sip.connected": "Registrado",
