+++ phoneblock-dongle/firmware/scripts/i18n/l10n/ui/ui_fr.arb
@@ -1,4 +1,5 @@
 {
+  "@@locale": "fr",
   "nav.home": "État",
   "status.title": "État",
   "status.sip.connected": "Enregistré",
