--- phoneblock-dongle/firmware/scripts/i18n/mail.de.json
@@ -1,41 +1,0 @@
-{
-  "intro": "Statusmeldung deines PhoneBlock-Dongles.",
-
-  "subj.error_and_calls": "PhoneBlock-Dongle: Fehler und neue Anrufe",
-  "subj.error": "PhoneBlock-Dongle: Fehler im Protokoll",
-  "subj.calls": "PhoneBlock-Dongle: Neue Anrufe",
-  "subj.status": "PhoneBlock-Dongle: Statusmeldung",
-  "subj.update": "PhoneBlock-Dongle: Firmware aktualisiert",
-
-  "newcalls.one": "Seit der letzten Meldung ist <b>1</b> neuer Anruf eingegangen.",
-  "newcalls.many": "Seit der letzten Meldung sind <b>%d</b> neue Anrufe eingegangen.",
-
-  "sum.device": "Ger&auml;t:",
-  "sum.uptime": "Laufzeit: %lldh %lldmin",
-  "sum.calls": "Anrufe gesamt: %u | SPAM blockiert: %u | durchgestellt: %u",
-
-  "calls.heading": "Letzte Anrufe",
-  "calls.time": "Zeit",
-  "calls.number": "Nummer",
-  "calls.name": "Name",
-  "calls.rating": "Bewertung",
-
-  "verdict.scope.range": "Bereich",
-  "verdict.scope.number": "Nummer",
-  "verdict.spam_blacklist": "SPAM (Blacklist, %s)",
-  "verdict.spam_blocklist": "SPAM (Blockliste, %s)",
-  "verdict.spam": "SPAM",
-  "verdict.spam_votes": "SPAM (%d direkt, %d Range)",
-  "verdict.spam_suspect": "SPAM-VERDACHT (%d direkt, %d Range)",
-  "verdict.legitimate": "legitim",
-  "verdict.error": "Fehler",
-  "verdict.unknown": "unbekannt",
-
-  "log.heading": "Neue Meldungen im Protokoll",
-
-  "update.intro1": "Die Firmware auf deinem ",
-  "update.link": "PhoneBlock-Dongle",
-  "update.intro2": " wurde auf ",
-  "update.version_prefix": "Version ",
-  "update.intro3": " aktualisiert."
-}
